requirements.txt
opencv-python
numpy
pillow
torch
torchvision
tk
pygame
